[fe] refactor (#78) 쿼리문 분리 - 쿼리를 담당하는 별도 파일 생성
